task-12 // second refact hw#12
