+++ 12-task/lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_flushbar/flutter_flushbar.dart';
-import 'screens/splashScreen.dart';
+import 'package:copy_task12/screens/splashScreen.dart';
 
 
 
@@ -26,22 +25,7 @@ class MyApp extends StatelessWidget {
   }
 }
 
-// функция показа сообщения об успешном копировании hex
-void showHexCopiedMessage(BuildContext context) {
-  Flushbar(
-    flushbarPosition: FlushbarPosition.BOTTOM,
-    backgroundColor: Colors.white,
-    duration: const Duration(seconds: 3),
-    messageText: const Center(
-      child: Text(
-        "Hex скопирован",
-        style: TextStyle(
-          fontSize: 20.0,
-        ),
-      ),
-    ),
-  ).show(context);
-}
+
