+++ 12-task/lib/model.dart
@@ -1,4 +1,3 @@
-// part of 'main.dart';
 import 'dart:convert';
 
 const String colors = """
@@ -65,18 +64,18 @@ const String colors = """
 }
 """;
 
-//декодирование джейсона и создание словаря {цвет: значение}, цвета без значения в словарь не попадают
-class MapJson{
-  final Map<String, String> colorMap = {};
-  final decodedJson = json.decode(colors);
-
-  void receivingMap(){
-    for (final color in decodedJson['colors']) {
-      if (color.containsKey('value')) {
-        colorMap[color['name']] = color['value'];
-      }
-
-    }
-  }
-}
+// //декодирование джейсона и создание словаря {цвет: значение}, цвета без значения в словарь не попадают
+// class MapJson{
+//   final Map<String, String> colorMap = {};
+//   final decodedJson = json.decode(colors);
+//
+//   void receivingMap(){
+//     for (final color in decodedJson['colors']) {
+//       if (color.containsKey('value')) {
+//         colorMap[color['name']] = color['value'];
+//       }
+//
+//     }
+//   }
+// }
